requirements.txt
django
django-crispy-forms
whitenoise
djangorestframework